espati laravel log

+++ app/Http/Controllers/emp_datfiscalesController.php
@@ -159,6 +159,11 @@ class emp_datfiscalesController extends AppBaseController
 
             return redirect(route('empDatfiscales.index'));
         }
+        if ($empDatfiscales->empresasacuerdo->count()>0)
+        {
+            Flash::error('Empresa no puede ser eliminada, tiene Acuerdos Comerciales Activos');
+            return redirect(route('empDatfiscales.index'));
+        }
 
         $this->empDatfiscalesRepository->delete($id);
